requirements.txt
altgraph==0.17.4
asgiref==3.10.0
contourpy==1.3.2
cycler==0.12.1
Django==5.0
djangorestframework==3.16.1
flatcam==0.8.5
fonttools==4.58.5
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pygame==2.6.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
rtree==1.4.0
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
svg.path==7.0
tzdata==2025.2
